requirements.txt
asgiref==3.8.1
Babel==2.15.0
black==24.4.2
click==8.1.7
Django==5.0.6
django-allauth==0.62.1
django-appconf==1.0.6
django-compressor==4.4
django-filter==24.2
django-money==3.5.2
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.1
py-moneyed==3.0
python-dotenv==1.0.1
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
